+++ frontend/controllers/shared/sidebar_controller.js
@@ -1,0 +1,13 @@
+import { Controller } from "stimulus"
+
+export default class extends Controller {
+  static targets = ["window"]
+
+  open() {
+    this.windowTarget.classList.add("open")
+  }
+
+  close() {
+    this.windowTarget.classList.remove("open")
+  }
+}